% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_functions.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download RedCap report data}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{key in .Renviron file to find redCap API token}

\item{redcapUrl}{URL to pull redCap data from}

\item{redcapReportId}{ID of redCap report to pull data from}
}
\value{
tibble of data contained at the redCap URL under the specified report ID
}
\description{
Given a redCap URL and report ID, downloads the report data using the users'
RedCap API token specified in .Renviron under the key redcapTokenName
}
\examples{
url <- "https://redcap.test.edu/api/"
reportId <- '12345'
tokenName <- 'mySecretAPItoken'
data = downloadRedcapReport(tokenName,url,reportId)
}
