% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_functions.R
\name{minimumN}
\alias{minimumN}
\title{Find minimum sample size}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{one sample of data (required)}

\item{x2}{second sample of data (option)}
}
\value{
minimum sample size for one-sample or two-sample t-test
}
\description{
Given 1 sample, x1, find the minimum sample size to determine if the mean of x1 differs from
0 with 80\% power with alpha = 0.05
}
\details{
Given 2 samples, x1 and x2, finds the minimum sample sizes to test if the means differ
}
\examples{
# One-sample t-test example
x1 = c(0, -1, -3, -6, 1, 2, -5)
minimumN(x1)

# Two-sample t-test example
x2 = c(10, 11, -1, 12, 0, 3, 15)
minimumN(x1, x2)
}
