% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_functions.R
\name{survCurv}
\alias{survCurv}
\title{Plot survival curve of time to event data}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{set of events (0/1 valued)}

\item{time}{time of each event}
}
\value{
ggplot object containing the survival curve
}
\description{
Given a set of events, status, with 1 corresponding to the event and 0 corresponding to a censored patient,
and the corresponding event times, time, plots a survival curve
}
\examples{
events = c(1,1,0,0,1,1,1,0,1,0)
year = c(2,3,4,5,7,9,9,10,12,15)
survCurv(events, year)
}
